feat(proto): apply authoritative proto refresh + add drift-check workflow
Regenerated RustPlusContracts.proto from the decompiled Rust dedicated
server (build 23601104) via tools/update-proto and applied the resulting
genuine wire changes to the typed API:

- New fields: AppCameraRays.time_of_day / camera_position / camera_rotation,
  ClanInfo.score, ClanInfo.Role.can_access_score_events,
  AppMarker.SellOrder.price_multiplier
- New marker type AppMarkerType.TravellingVendor (+ TravellingVendorMarker)
- Renames/fixes: AppMap.Monument.name -> token, AppMarkerType.Unknow -> Undefined
- Entity-id widening uint32 -> uint64 across markers, camera entities,
  entity-changed events and RustPlus/IRustPlus params (NetworkableId is 64-bit)
- AppTeamInfo.Note adopts server truth (int32 icon / colour_index / label);
  dropped the redundant IconType/IconColor proto enums (the library keeps its
  own NoteIcons/NoteColors)

The committed proto now reproduces byte-for-byte from a fresh server download
(update-proto.sh exits clean). Adds .github/workflows/ProtoRefresh.yml: a
first-Thursday + on-demand workflow that reruns the pipeline and opens/updates
a proto-drift issue with the diff when the server changes (never auto-applies).

Build clean (0 warnings), 51 tests green.

+++ src/RustPlusApi/Data/Cameras/CameraEntity.cs
@@ -4,7 +4,7 @@ namespace RustPlusApi.Data.Cameras;
 public sealed record CameraEntity
 {
     /// <summary>Unique entity identifier.</summary>
-    public uint EntityId { get; init; }
+    public ulong EntityId { get; init; }
 
     /// <summary>The kind of entity (player, tree, etc.).</summary>
     public CameraEntityType Type { get; init; }
